@@ -28,11 +28,12 @@ export class ColorWheel extends Component {
     }
   }
 
-  UNSAFE_componentWillMount = () => {
+  componentDidMount = () => {
     this._panResponder = PanResponder.create({
       onStartShouldSetPanResponderCapture: ({nativeEvent}) => {
         if (this.outBounds(nativeEvent)) return
         this.updateColor({nativeEvent})
+        this.setState({panHandlerReady: true})
 
         this.state.pan.setValue({
           x: -this.state.left + nativeEvent.pageX - this.props.thumbSize / 2,
@@ -44,18 +45,23 @@ export class ColorWheel extends Component {
       onMoveShouldSetPanResponderCapture: () => true,
       onPanResponderGrant: () => true,
       onPanResponderMove: (event, gestureState) => {
-        const {deg, radius} = this.calcPolar(gestureState);
-        const {left, top} = this.calcCartesian(deg, (radius > 1) ? 1 : radius);
+        if (this.outBounds(gestureState)) return
 
-        this.state.pan.setValue({
-          x: left - this.props.thumbSize / 2,
-          y: top - this.props.thumbSize / 2,
-        })
-
-        this.updateColor(event, gestureState);
+        this.resetPanHandler()
+        return Animated.event(
+          [
+            null,
+            {
+              dx: this.state.pan.x,
+              dy: this.state.pan.y,
+            },
+          ],
+          {listener: this.updateColor}
+        )(event, gestureState)
       },
       onMoveShouldSetPanResponder: () => true,
       onPanResponderRelease: ({nativeEvent}) => {
+        this.setState({panHandlerReady: true})
         this.state.pan.flattenOffset()
         const {radius} = this.calcPolar(nativeEvent)
         if (radius < 0.1) {
@@ -115,6 +121,19 @@ export class ColorWheel extends Component {
     return radius > 1
   }
 
+  resetPanHandler () {
+    if (!this.state.panHandlerReady) {
+      return
+    }
+
+    this.setState({panHandlerReady: false})
+    this.state.pan.setOffset({
+      x: this.state.pan.x._value,
+      y: this.state.pan.y._value,
+    })
+    this.state.pan.setValue({x: 0, y: 0})
+  }
+
   calcCartesian (deg, radius) {
     const r = radius * this.state.radius // was normalized
     const rad = Math.PI * deg / 180
@@ -128,9 +147,9 @@ export class ColorWheel extends Component {
 
   updateColor = ({nativeEvent}) => {
     const {deg, radius} = this.calcPolar(nativeEvent)
-    const currentColor = colorsys.hsv2Hex({h: deg, s: 100 * ((radius < 1) ? radius : 1), v: 100})
+    const currentColor = colorsys.hsv2Hex({h: deg, s: 100 * radius, v: 100})
     this.setState({currentColor})
-    this.props.onColorChange({h: deg, s: 100 * ((radius < 1) ? radius : 1), v: 100})
+    this.props.onColorChange({h: deg, s: 100 * radius, v: 100})
   }
 
   forceUpdate = color => {
@@ -171,17 +190,22 @@ export class ColorWheel extends Component {
       },
     ]
 
+    const panHandlers = this._panResponder && this._panResponder.panHandlers || {}
+
     return (
       <View
         ref={node => {
           this.self = node
         }}
-        {...this._panResponder.panHandlers}
+        {...panHandlers}
         onLayout={nativeEvent => this.onLayout(nativeEvent)}
-        style={[styles.coverResponder, this.props.style]}
-      >
+        style={[styles.coverResponder, this.props.style]}>
         <Image
-          style={[styles.img, {height: radius * 2, width: radius * 2}]}
+          style={[styles.img, 
+                  {
+                    height: radius * 2 - this.props.thumbSize,
+                    width: radius * 2 - this.props.thumbSize
+                  }]}
           source={require('./color-wheel.png')}
         />
         <Animated.View style={[this.state.pan.getLayout(), thumbStyle]} />
